requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.12.1
filelock==3.0.12
idna==2.10
importlib-metadata==2.0.0
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.1
pytz==2020.1
requests==2.24.0
six==1.15.0
some-package==0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.0.33
wrapt==1.12.1
zipp==3.3.0
